Create Polymorphism2.cpp
Currently, the child class inherits from parent<int>, meaning T is equal to int. We can change the data type T and see what happens. Now, T is set to float. The Add function in the parent class adds two values, 5.5f and 10.5f, which results in 16.0f. Then, in the Add function of the child class, the value 20.0f is added to it, and the final result becomes 36.0f.
Output: Result: 36