Create Polymorphism3.cpp
Removing the virtual and override Keywords Impact on Polymorphism

Currently, the `Add` function is defined as `virtual` in the parent class and overridden with override in the child class. We can remove these keywords and see what happens.
Removing virtual and override
By removing virtual the Add function is no longer a virtual function, and run-time polymorphism is lost. If we use a pointer or reference of type parent<int> to call the Add function, the Add function from the parent class will be executed, not the one from the child class. However, in this specific code, because we explicitly call parent<int> Add the output does not change and remains 35.
Output: Result : 35
Removing virtual disables run-time polymorphism and causes the decision to be made at compile-time (Static Dispatch). This can impact performance removing virtual can make the program run faster because the virtual function table (vtable) is not created but it sacrifices the flexibility of polymorphism.