Create Polymorphism4.cpp
Changing the access of the Add function in the parent class from protected to public
Currently, the Add function in the parent class is defined as protected, meaning only the class itself and its derived classes can access it. We can change this to public. What happens?
Now, the Add function from the parent class is accessible from outside as well. For example, we can directly call the Add function from the parent class in the main function : 
parentObj.Add(3, 4)which results in 3 + 4 = 7. obj.Add(3, 4) gives the same previous result (35).
Why is this change important?
Changing the access level (access modifier) can affect how the class is used. By making the Add function public, you can directly use it from the parent class. However, this might violate encapsulation.